+++ Assets/script/game/Prefab/Store Card/StoreCardAppearance.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class StoreCardAppearance : MonoBehaviour
 {
     [SerializeField] private Renderer _renderer;
-    public void SetImage(byte cardIndex)
+    public void SetImage(int cardIndex)
     {
         Material material = Util.TransferStoreCardSprite(cardIndex);
         _renderer.material = material;
